+++ src/main/java/net/voidz/block/render/PortalBlockEntityRenderer.java
@@ -1,45 +1,38 @@
 package net.voidz.block.render;
 
-import org.joml.Matrix4f;
-
 import net.fabricmc.api.EnvType;
 import net.fabricmc.api.Environment;
-import net.minecraft.client.render.RenderLayer;
-import net.minecraft.client.render.VertexConsumer;
 import net.minecraft.client.render.VertexConsumerProvider;
-import net.minecraft.client.render.block.entity.BlockEntityRenderer;
 import net.minecraft.client.render.block.entity.BlockEntityRendererFactory;
+import net.minecraft.client.render.block.entity.EndPortalBlockEntityRenderer;
 import net.minecraft.client.util.math.MatrixStack;
-import net.minecraft.util.Identifier;
-import net.minecraft.util.math.Direction;
 import net.voidz.block.entity.PortalBlockEntity;
 
 @Environment(EnvType.CLIENT)
-public class PortalBlockEntityRenderer<T extends PortalBlockEntity> implements BlockEntityRenderer<T> {
-    public static final Identifier PORTAL_TEXTURE = new Identifier("textures/entity/end_portal.png");;
+public class PortalBlockEntityRenderer extends EndPortalBlockEntityRenderer<PortalBlockEntity> {
 
     public PortalBlockEntityRenderer(BlockEntityRendererFactory.Context ctx) {
+        super(ctx);
     }
 
     @Override
-    public void render(T portalBlockEntity, float f, MatrixStack matrixStack, VertexConsumerProvider vertexConsumerProvider, int i, int j) {
-        Matrix4f matrix4f = matrixStack.peek().getPositionMatrix();
-        this.renderSides(portalBlockEntity, matrix4f, vertexConsumerProvider.getBuffer(RenderLayer.getEndPortal()));
+    public void render(PortalBlockEntity endPortalBlockEntity, float f, MatrixStack matrixStack, VertexConsumerProvider vertexConsumerProvider, int i, int j) {
+        super.render(endPortalBlockEntity, f, matrixStack, vertexConsumerProvider, i, j);
     }
 
-    private void renderSides(T entity, Matrix4f matrix4f, VertexConsumer vertexConsumer) {
-        this.renderSide(entity, matrix4f, vertexConsumer, 0.0F, 1.0F, 0.0F, 1.0F, 1.0F, 1.0F, 1.0F, 1.0F, Direction.SOUTH);
-        this.renderSide(entity, matrix4f, vertexConsumer, 0.0F, 1.0F, 1.0F, 0.0F, 0.0F, 0.0F, 0.0F, 0.0F, Direction.NORTH);
-        this.renderSide(entity, matrix4f, vertexConsumer, 1.0F, 1.0F, 1.0F, 0.0F, 0.0F, 1.0F, 1.0F, 0.0F, Direction.EAST);
-        this.renderSide(entity, matrix4f, vertexConsumer, 0.0F, 0.0F, 0.0F, 1.0F, 0.0F, 1.0F, 1.0F, 0.0F, Direction.WEST);
-        this.renderSide(entity, matrix4f, vertexConsumer, 0.0F, 1.0F, 0.0F, 0.0F, 0.0F, 0.0F, 1.0F, 1.0F, Direction.DOWN);
-        this.renderSide(entity, matrix4f, vertexConsumer, 0.0F, 1.0F, 1.0F, 1.0F, 1.0F, 1.0F, 0.0F, 0.0F, Direction.UP);
+    @Override
+    protected float getTopYOffset() {
+        return 1.0f;
     }
 
-    private void renderSide(T entity, Matrix4f model, VertexConsumer vertices, float x1, float x2, float y1, float y2, float z1, float z2, float z3, float z4, Direction direction) {
-        vertices.vertex(model, x1, y1, z1).next();
-        vertices.vertex(model, x2, y1, z2).next();
-        vertices.vertex(model, x2, y2, z3).next();
-        vertices.vertex(model, x1, y2, z4).next();
+    @Override
+    protected float getBottomYOffset() {
+        return 0.0f;
     }
+
+    @Override
+    public int getRenderDistance() {
+        return 256;
+    }
+
 }