+++ CHANGELOG.md
@@ -3,5 +3,4 @@
 ### Fixed:
 - 
 ### Changed:
-- Changed portal block to solid
-- Changed void stone block name
+- Changed portal block renderer